requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
alembic==1.18.4
annotated-types==0.7.0
attrs==25.4.0
certifi==2026.2.25
charset-normalizer==3.4.4
frozenlist==1.8.0
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
multidict==6.7.1
propcache==0.4.1
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
PyJWT==2.11.0
python-dotenv==1.2.2
requests==2.32.5
SQLAlchemy==2.0.48
twilio==9.10.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
yarl==1.23.0
